move royalty fee tests to a new suite class
