+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/suites/crypto/TransferWithCustomFixedFees.java
@@ -29,14 +29,9 @@ import static com.hedera.services.bdd.spec.transactions.TxnVerbs.mintToken;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.tokenAssociate;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.tokenCreate;
 import static com.hedera.services.bdd.spec.transactions.token.CustomFeeSpecs.fixedHbarFee;
-import static com.hedera.services.bdd.spec.transactions.token.CustomFeeSpecs.fixedHbarFeeInheritingRoyaltyCollector;
 import static com.hedera.services.bdd.spec.transactions.token.CustomFeeSpecs.fixedHtsFee;
-import static com.hedera.services.bdd.spec.transactions.token.CustomFeeSpecs.fixedHtsFeeInheritingRoyaltyCollector;
 import static com.hedera.services.bdd.spec.transactions.token.CustomFeeSpecs.fractionalFee;
-import static com.hedera.services.bdd.spec.transactions.token.CustomFeeSpecs.royaltyFeeNoFallback;
-import static com.hedera.services.bdd.spec.transactions.token.CustomFeeSpecs.royaltyFeeWithFallback;
 import static com.hedera.services.bdd.spec.transactions.token.TokenMovement.moving;
-import static com.hedera.services.bdd.spec.transactions.token.TokenMovement.movingHbar;
 import static com.hedera.services.bdd.spec.transactions.token.TokenMovement.movingUnique;
 import static com.hedera.services.bdd.spec.transactions.token.TokenMovement.movingUniqueWithAllowance;
 import static com.hedera.services.bdd.spec.transactions.token.TokenMovement.movingWithAllowance;
@@ -44,7 +39,6 @@ import static com.hedera.services.bdd.spec.utilops.CustomSpecAssert.allRunFor;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.newKeyNamed;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.withOpContext;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.*;
-
 import com.hedera.node.app.hapi.utils.ByteStringUtils;
 import com.hedera.services.bdd.junit.HapiTest;
 import com.hedera.services.bdd.junit.HapiTestSuite;
@@ -1590,311 +1584,6 @@ public class TransferWithCustomFixedFees extends HapiSuite {
                         .fee(ONE_HUNDRED_HBARS)
                         .payingWith(tokenOwner)
                         .hasKnownStatus(INSUFFICIENT_SENDER_ACCOUNT_BALANCE_FOR_CUSTOM_FEE));
-    }
-
-    @HapiTest
-    public HapiSpec transferNonFungibleWithRoyaltyHbarFee() {
-        final var fourHundredHbars = 4 * ONE_HUNDRED_HBARS;
-        final var twoHundredHbars = 2 * ONE_HUNDRED_HBARS;
-        return defaultHapiSpec("transferNonFungibleWithRoyaltyHbarFee")
-                .given(
-                        newKeyNamed(NFT_KEY),
-                        cryptoCreate(hbarCollector).balance(0L),
-                        cryptoCreate(tokenReceiver).balance(ONE_MILLION_HBARS),
-                        cryptoCreate(tokenTreasury),
-                        cryptoCreate(tokenOwner).balance(0L),
-                        tokenCreate(nonFungibleToken)
-                                .treasury(tokenTreasury)
-                                .tokenType(TokenType.NON_FUNGIBLE_UNIQUE)
-                                .initialSupply(0)
-                                .supplyKey(NFT_KEY)
-                                .supplyType(TokenSupplyType.INFINITE)
-                                .withCustom(royaltyFeeNoFallback(1, 2, hbarCollector)),
-                        tokenAssociate(tokenReceiver, nonFungibleToken),
-                        tokenAssociate(tokenOwner, nonFungibleToken),
-                        mintToken(nonFungibleToken, List.of(ByteStringUtils.wrapUnsafely("meta1".getBytes()))),
-                        cryptoTransfer(movingUnique(nonFungibleToken, 1L).between(tokenTreasury, tokenOwner)))
-                .when(cryptoTransfer(
-                        movingUnique(nonFungibleToken, 1L).between(tokenOwner, tokenReceiver),
-                        movingHbar(fourHundredHbars).between(tokenReceiver, tokenOwner)))
-                .then(
-                        getAccountBalance(tokenOwner)
-                                .hasTinyBars(twoHundredHbars)
-                                .hasTokenBalance(nonFungibleToken, 0),
-                        getAccountBalance(tokenReceiver).hasTokenBalance(nonFungibleToken, 1),
-                        getAccountBalance(hbarCollector).hasTinyBars(twoHundredHbars));
-    }
-
-    @HapiTest
-    public HapiSpec transferNonFungibleWithRoyaltyFungibleFee() {
-        return defaultHapiSpec("transferNonFungibleWithRoyaltyFungibleFee")
-                .given(
-                        newKeyNamed(NFT_KEY),
-                        cryptoCreate(htsCollector),
-                        cryptoCreate(tokenReceiver).balance(ONE_MILLION_HBARS),
-                        cryptoCreate(tokenTreasury),
-                        cryptoCreate(tokenOwner),
-                        tokenCreate(feeDenom).treasury(tokenReceiver).initialSupply(4),
-                        tokenAssociate(tokenOwner, feeDenom),
-                        tokenAssociate(htsCollector, feeDenom),
-                        tokenCreate(nonFungibleToken)
-                                .treasury(tokenTreasury)
-                                .tokenType(TokenType.NON_FUNGIBLE_UNIQUE)
-                                .initialSupply(0)
-                                .supplyKey(NFT_KEY)
-                                .supplyType(TokenSupplyType.INFINITE)
-                                .withCustom(royaltyFeeNoFallback(1, 2, htsCollector)),
-                        tokenAssociate(tokenReceiver, nonFungibleToken),
-                        tokenAssociate(tokenOwner, nonFungibleToken),
-                        mintToken(nonFungibleToken, List.of(ByteStringUtils.wrapUnsafely("meta1".getBytes()))),
-                        cryptoTransfer(movingUnique(nonFungibleToken, 1L).between(tokenTreasury, tokenOwner)))
-                .when(cryptoTransfer(
-                        movingUnique(nonFungibleToken, 1L).between(tokenOwner, tokenReceiver),
-                        moving(4, feeDenom).between(tokenReceiver, tokenOwner)))
-                .then(
-                        getAccountBalance(tokenOwner).hasTokenBalance(nonFungibleToken, 0).hasTokenBalance(feeDenom, 2),
-                        getAccountBalance(tokenReceiver)
-                                .hasTokenBalance(nonFungibleToken, 1)
-                                .hasTokenBalance(feeDenom, 0),
-                        getAccountBalance(htsCollector).hasTokenBalance(feeDenom, 2));
-    }
-
-    @HapiTest
-    public HapiSpec transferNonFungibleWithRoyaltyFallbackHbarFee() {
-        return defaultHapiSpec("transferNonFungibleWithRoyaltyFallbackHbarFee")
-                .given(
-                        newKeyNamed(NFT_KEY),
-                        cryptoCreate(hbarCollector).balance(0L),
-                        cryptoCreate(tokenReceiver).balance(ONE_MILLION_HBARS),
-                        cryptoCreate(tokenTreasury),
-                        cryptoCreate(tokenOwner),
-                        tokenCreate(nonFungibleToken)
-                                .treasury(tokenTreasury)
-                                .tokenType(TokenType.NON_FUNGIBLE_UNIQUE)
-                                .initialSupply(0)
-                                .supplyKey(NFT_KEY)
-                                .supplyType(TokenSupplyType.INFINITE)
-                                .withCustom(royaltyFeeWithFallback(
-                                        1, 2, fixedHbarFeeInheritingRoyaltyCollector(100), hbarCollector)),
-                        tokenAssociate(tokenReceiver, nonFungibleToken),
-                        tokenAssociate(tokenOwner, nonFungibleToken),
-                        mintToken(nonFungibleToken, List.of(ByteStringUtils.wrapUnsafely("meta1".getBytes()))),
-                        cryptoTransfer(movingUnique(nonFungibleToken, 1L).between(tokenTreasury, tokenOwner)))
-                .when(cryptoTransfer(movingUnique(nonFungibleToken, 1L).between(tokenOwner, tokenReceiver))
-                        .signedByPayerAnd(tokenReceiver, tokenOwner))
-                .then(
-                        getAccountBalance(tokenOwner).hasTokenBalance(nonFungibleToken, 0),
-                        getAccountBalance(tokenReceiver)
-                                .hasTokenBalance(nonFungibleToken, 1)
-                                .hasTinyBars(ONE_MILLION_HBARS - 100),
-                        getAccountBalance(hbarCollector).hasTinyBars(100));
-    }
-
-    @HapiTest
-    public HapiSpec transferNonFungibleWithRoyaltyFallbackFungibleFee() {
-        return defaultHapiSpec("transferNonFungibleWithRoyaltyFallbackFungibleFee")
-                .given(
-                        newKeyNamed(NFT_KEY),
-                        cryptoCreate(htsCollector),
-                        cryptoCreate(tokenReceiver).balance(ONE_MILLION_HBARS),
-                        cryptoCreate(tokenTreasury),
-                        cryptoCreate(tokenOwner),
-                        tokenCreate(feeDenom).treasury(tokenReceiver).initialSupply(5),
-                        tokenAssociate(tokenOwner, feeDenom),
-                        tokenAssociate(htsCollector, feeDenom),
-                        tokenCreate(nonFungibleToken)
-                                .treasury(tokenTreasury)
-                                .tokenType(TokenType.NON_FUNGIBLE_UNIQUE)
-                                .initialSupply(0)
-                                .supplyKey(NFT_KEY)
-                                .supplyType(TokenSupplyType.INFINITE)
-                                .withCustom(royaltyFeeWithFallback(
-                                        1, 2, fixedHtsFeeInheritingRoyaltyCollector(5, feeDenom), htsCollector)),
-                        tokenAssociate(tokenReceiver, nonFungibleToken),
-                        tokenAssociate(tokenOwner, nonFungibleToken),
-                        mintToken(nonFungibleToken, List.of(ByteStringUtils.wrapUnsafely("meta1".getBytes()))),
-                        cryptoTransfer(movingUnique(nonFungibleToken, 1L).between(tokenTreasury, tokenOwner)))
-                .when(cryptoTransfer(movingUnique(nonFungibleToken, 1L).between(tokenOwner, tokenReceiver))
-                        .signedByPayerAnd(tokenReceiver, tokenOwner))
-                .then(
-                        getAccountBalance(tokenOwner).hasTokenBalance(nonFungibleToken, 0),
-                        getAccountBalance(tokenReceiver)
-                                .hasTokenBalance(nonFungibleToken, 1)
-                                .hasTokenBalance(feeDenom, 0),
-                        getAccountBalance(htsCollector).hasTokenBalance(feeDenom, 5));
-    }
-
-    @HapiTest
-    public HapiSpec transferNonFungibleWithRoyaltyHbarFeeInsufficientBalance() {
-        final var fourHundredHbars = 4 * ONE_HUNDRED_HBARS;
-        return defaultHapiSpec("transferNonFungibleWithRoyaltyHbarFeeInsufficientBalance")
-                .given(
-                        newKeyNamed(NFT_KEY),
-                        cryptoCreate(hbarCollector).balance(0L),
-                        cryptoCreate(tokenReceiver).balance(0L),
-                        cryptoCreate(tokenTreasury),
-                        cryptoCreate(tokenOwner).balance(0L),
-                        tokenCreate(nonFungibleToken)
-                                .treasury(tokenTreasury)
-                                .tokenType(TokenType.NON_FUNGIBLE_UNIQUE)
-                                .initialSupply(0)
-                                .supplyKey(NFT_KEY)
-                                .supplyType(TokenSupplyType.INFINITE)
-                                .withCustom(royaltyFeeNoFallback(1, 2, hbarCollector)),
-                        tokenAssociate(tokenReceiver, nonFungibleToken),
-                        tokenAssociate(tokenOwner, nonFungibleToken),
-                        mintToken(nonFungibleToken, List.of(ByteStringUtils.wrapUnsafely("meta1".getBytes()))),
-                        cryptoTransfer(movingUnique(nonFungibleToken, 1L).between(tokenTreasury, tokenOwner)))
-                .when()
-                .then(cryptoTransfer(
-                                movingUnique(nonFungibleToken, 1L).between(tokenOwner, tokenReceiver),
-                                movingHbar(fourHundredHbars).between(tokenReceiver, tokenOwner))
-                        .hasKnownStatus(INSUFFICIENT_ACCOUNT_BALANCE));
-    }
-
-    @HapiTest
-    public HapiSpec transferNonFungibleWithRoyaltyFungibleFeeInsufficientBalance() {
-        return defaultHapiSpec("transferNonFungibleWithRoyaltyFungibleFeeInsufficientBalance")
-                .given(
-                        newKeyNamed(NFT_KEY),
-                        cryptoCreate(htsCollector),
-                        cryptoCreate(tokenReceiver).balance(ONE_MILLION_HBARS),
-                        cryptoCreate(tokenTreasury),
-                        cryptoCreate(tokenOwner),
-                        tokenCreate(feeDenom).treasury(tokenReceiver).initialSupply(0),
-                        tokenAssociate(tokenOwner, feeDenom),
-                        tokenAssociate(htsCollector, feeDenom),
-                        tokenCreate(nonFungibleToken)
-                                .treasury(tokenTreasury)
-                                .tokenType(TokenType.NON_FUNGIBLE_UNIQUE)
-                                .initialSupply(0)
-                                .supplyKey(NFT_KEY)
-                                .supplyType(TokenSupplyType.INFINITE)
-                                .withCustom(royaltyFeeNoFallback(1, 2, htsCollector)),
-                        tokenAssociate(tokenReceiver, nonFungibleToken),
-                        tokenAssociate(tokenOwner, nonFungibleToken),
-                        mintToken(nonFungibleToken, List.of(ByteStringUtils.wrapUnsafely("meta1".getBytes()))),
-                        cryptoTransfer(movingUnique(nonFungibleToken, 1L).between(tokenTreasury, tokenOwner)))
-                .when()
-                .then(cryptoTransfer(
-                                movingUnique(nonFungibleToken, 1L).between(tokenOwner, tokenReceiver),
-                                moving(4, feeDenom).between(tokenReceiver, tokenOwner))
-                        .hasKnownStatus(INSUFFICIENT_TOKEN_BALANCE));
-    }
-
-    @HapiTest
-    public HapiSpec transferNonFungibleWithRoyaltyFallbackHbarFeeInsufficientBalance() {
-        return defaultHapiSpec("transferNonFungibleWithRoyaltyFallbackHbarFeeInsufficientBalance")
-                .given(
-                        newKeyNamed(NFT_KEY),
-                        cryptoCreate(hbarCollector).balance(0L),
-                        cryptoCreate(tokenReceiver).balance(0L),
-                        cryptoCreate(tokenTreasury),
-                        cryptoCreate(tokenOwner),
-                        tokenCreate(nonFungibleToken)
-                                .treasury(tokenTreasury)
-                                .tokenType(TokenType.NON_FUNGIBLE_UNIQUE)
-                                .initialSupply(0)
-                                .supplyKey(NFT_KEY)
-                                .supplyType(TokenSupplyType.INFINITE)
-                                .withCustom(royaltyFeeWithFallback(
-                                        1, 2, fixedHbarFeeInheritingRoyaltyCollector(100), hbarCollector)),
-                        tokenAssociate(tokenReceiver, nonFungibleToken),
-                        tokenAssociate(tokenOwner, nonFungibleToken),
-                        mintToken(nonFungibleToken, List.of(ByteStringUtils.wrapUnsafely("meta1".getBytes()))),
-                        cryptoTransfer(movingUnique(nonFungibleToken, 1L).between(tokenTreasury, tokenOwner)))
-                .when()
-                .then(cryptoTransfer(movingUnique(nonFungibleToken, 1L).between(tokenOwner, tokenReceiver))
-                        .signedByPayerAnd(tokenReceiver, tokenOwner)
-                        .hasKnownStatus(INSUFFICIENT_ACCOUNT_BALANCE));
-    }
-
-    @HapiTest
-    public HapiSpec transferNonFungibleWithRoyaltyFallbackFungibleFeeInsufficientBalance() {
-        return defaultHapiSpec("transferNonFungibleWithRoyaltyFallbackFungibleFeeInsufficientBalance")
-                .given(
-                        newKeyNamed(NFT_KEY),
-                        cryptoCreate(htsCollector),
-                        cryptoCreate(tokenReceiver).balance(ONE_MILLION_HBARS),
-                        cryptoCreate(tokenTreasury),
-                        cryptoCreate(tokenOwner),
-                        tokenCreate(feeDenom).treasury(tokenReceiver).initialSupply(0),
-                        tokenAssociate(tokenOwner, feeDenom),
-                        tokenAssociate(htsCollector, feeDenom),
-                        tokenCreate(nonFungibleToken)
-                                .treasury(tokenTreasury)
-                                .tokenType(TokenType.NON_FUNGIBLE_UNIQUE)
-                                .initialSupply(0)
-                                .supplyKey(NFT_KEY)
-                                .supplyType(TokenSupplyType.INFINITE)
-                                .withCustom(royaltyFeeWithFallback(
-                                        1, 2, fixedHtsFeeInheritingRoyaltyCollector(5, feeDenom), htsCollector)),
-                        tokenAssociate(tokenReceiver, nonFungibleToken),
-                        tokenAssociate(tokenOwner, nonFungibleToken),
-                        mintToken(nonFungibleToken, List.of(ByteStringUtils.wrapUnsafely("meta1".getBytes()))),
-                        cryptoTransfer(movingUnique(nonFungibleToken, 1L).between(tokenTreasury, tokenOwner)))
-                .when()
-                .then(cryptoTransfer(movingUnique(nonFungibleToken, 1L).between(tokenOwner, tokenReceiver))
-                        .signedByPayerAnd(tokenReceiver, tokenOwner)
-                        .hasKnownStatus(INSUFFICIENT_SENDER_ACCOUNT_BALANCE_FOR_CUSTOM_FEE));
-    }
-
-    @HapiTest
-    public HapiSpec transferNonFungibleWithRoyaltyFallbackFungibleFeeNoAssociation() {
-        return defaultHapiSpec("transferNonFungibleWithRoyaltyFallbackFungibleFeeNoAssociation")
-                .given(
-                        newKeyNamed(NFT_KEY),
-                        cryptoCreate(htsCollector),
-                        cryptoCreate(tokenReceiver).balance(ONE_MILLION_HBARS),
-                        cryptoCreate(tokenTreasury),
-                        cryptoCreate(tokenOwner),
-                        tokenCreate(feeDenom).treasury(tokenTreasury).initialSupply(0),
-                        tokenAssociate(htsCollector, feeDenom),
-                        tokenCreate(nonFungibleToken)
-                                .treasury(tokenTreasury)
-                                .tokenType(TokenType.NON_FUNGIBLE_UNIQUE)
-                                .initialSupply(0)
-                                .supplyKey(NFT_KEY)
-                                .supplyType(TokenSupplyType.INFINITE)
-                                .withCustom(royaltyFeeWithFallback(
-                                        1, 2, fixedHtsFeeInheritingRoyaltyCollector(5, feeDenom), htsCollector)),
-                        tokenAssociate(tokenReceiver, nonFungibleToken),
-                        tokenAssociate(tokenOwner, nonFungibleToken),
-                        mintToken(nonFungibleToken, List.of(ByteStringUtils.wrapUnsafely("meta1".getBytes()))),
-                        cryptoTransfer(movingUnique(nonFungibleToken, 1L).between(tokenTreasury, tokenOwner)))
-                .when()
-                .then(cryptoTransfer(movingUnique(nonFungibleToken, 1L).between(tokenOwner, tokenReceiver))
-                        .signedByPayerAnd(tokenReceiver, tokenOwner)
-                        .hasKnownStatus(TOKEN_NOT_ASSOCIATED_TO_ACCOUNT));
-    }
-
-    @HapiTest
-    public HapiSpec transferNonFungibleWithRoyaltyFungibleFeeNoAssociation() {
-        return defaultHapiSpec("transferNonFungibleWithRoyaltyFungibleFeeNoAssociation")
-                .given(
-                        newKeyNamed(NFT_KEY),
-                        cryptoCreate(htsCollector),
-                        cryptoCreate(tokenReceiver).balance(ONE_MILLION_HBARS),
-                        cryptoCreate(tokenTreasury),
-                        cryptoCreate(tokenOwner),
-                        tokenCreate(feeDenom).treasury(tokenTreasury).initialSupply(0),
-                        tokenAssociate(htsCollector, feeDenom),
-                        tokenCreate(nonFungibleToken)
-                                .treasury(tokenTreasury)
-                                .tokenType(TokenType.NON_FUNGIBLE_UNIQUE)
-                                .initialSupply(0)
-                                .supplyKey(NFT_KEY)
-                                .supplyType(TokenSupplyType.INFINITE)
-                                .withCustom(royaltyFeeNoFallback(1, 2, htsCollector)),
-                        tokenAssociate(tokenReceiver, nonFungibleToken),
-                        tokenAssociate(tokenOwner, nonFungibleToken),
-                        mintToken(nonFungibleToken, List.of(ByteStringUtils.wrapUnsafely("meta1".getBytes()))),
-                        cryptoTransfer(movingUnique(nonFungibleToken, 1L).between(tokenTreasury, tokenOwner)))
-                .when()
-                .then(cryptoTransfer(
-                                movingUnique(nonFungibleToken, 1L).between(tokenOwner, tokenReceiver),
-                                moving(4, feeDenom).between(tokenReceiver, tokenOwner))
-                        .hasKnownStatus(TOKEN_NOT_ASSOCIATED_TO_ACCOUNT));
     }
 
     @Override
